Experiment finding TD templates for Pod coverage test.
This kind of works, but because Pod::Coverage by default ignores imported
functions, it misses wrappers created by `create_wrapper`. Furthermore,
because we're changing the funciton name back to its original template name
with this hack, it also finds imported/aliased templates. So with this code
uncommented, I get:

    not ok 4 - Pod coverage on MyApp::Templates::HTML

    #   Failed test 'Pod coverage on MyApp::Templates::HTML'
    #   at /usr/local/lib/perl5/site_perl/5.10.1/Test/Pod/Coverage.pm line 126.
    # Coverage for MyApp::Templates::HTML is 0.0%, with 1 naked subroutine:
    # 	books/list

Which is not okay. I think that we'd really need to create a separate test
module to ensure that templates are documented, which is unfortunate. But I
don't really see another way around it at this point.

Anyway, experiment done for now. Moving on.